requirements.txt
Django==1.5.1
django-appconf==0.6
django-floppyforms==1.1
django-crispy-forms
django-model-utils
six==1.3.0
wsgiref==0.1.2
django-jsonfield
django-autoslug
